check if user is moderator for admin level requests

+++ src/main/java/com/inix/omqweb/BeatmapReport/BeatmapReportController.java
@@ -1,7 +1,13 @@
 package com.inix.omqweb.BeatmapReport;
 
+import com.inix.omqweb.Achievement.AchievementService;
 import com.inix.omqweb.Beatmap.BeatmapService;
+import com.inix.omqweb.osuAPI.Player;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,6 +18,9 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BeatmapReportController {
     private final BeatmapService beatmapService;
+    private final AchievementService achievementService;
+
+    private final Logger logger = LoggerFactory.getLogger(BeatmapReportController.class);
 
     @PostMapping("/reportBeatmap")
     public void reportBeatmap(@RequestBody BeatmapReportDTO beatmapReportDTO) {
@@ -19,17 +28,40 @@ public class BeatmapReportController {
     }
 
     @GetMapping("/beatmapReports")
-    public ResponseEntity<List<BeatmapReportResponseDTO>> getBeatmapReports() {
-        return ResponseEntity.ok(beatmapService.getMostReportedBeatmaps());
+    public ResponseEntity<List<BeatmapReportResponseDTO>> getBeatmapReports(HttpServletRequest request) {
+        Player player = (Player) request.getAttribute("userInfo");
+
+        if (achievementService.isModerator(player)) {
+            return ResponseEntity.ok(beatmapService.getMostReportedBeatmaps());
+        } else {
+            logger.warn("Player {} tried to get beatmap reports", player.getId());
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @PostMapping("/addBlur")
-    public void addBlur(@RequestParam String beatmapsetId) {
-        beatmapService.addBlur(beatmapsetId);
+    public ResponseEntity<?> addBlur(@RequestParam String beatmapsetId, HttpServletRequest request) {
+        Player player = (Player) request.getAttribute("userInfo");
+
+        if (achievementService.isModerator(player)) {
+            beatmapService.addBlur(beatmapsetId);
+            return ResponseEntity.ok().build();
+        } else {
+            logger.warn("Player {} tried to add blur to beatmapset {}", player.getId(), beatmapsetId);
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @PostMapping("/deleteReport")
-    public void deleteReport(@RequestParam String beatmapsetId) {
-        beatmapService.deleteReport(beatmapsetId);
+    public ResponseEntity<?> deleteReport(@RequestParam String beatmapsetId, HttpServletRequest request) {
+        Player player = (Player) request.getAttribute("userInfo");
+
+        if (achievementService.isModerator(player)) {
+            beatmapService.deleteReport(beatmapsetId);
+            return ResponseEntity.ok().build();
+        } else {
+            logger.warn("Player {} tried to delete report from beatmapset {}", player.getId(), beatmapsetId);
+            return ResponseEntity.badRequest().build();
+        }
     }
 }
